requirements.txt
fastapi[standard]
motor
kafka-python
aioredis
concurrent_log_handler
aiohappyeyeballs
aiohttp
aioredis
aiosignal
annotated-types
anthropic
anyio
beautifulsoup4
blinker
concurrent-log-handler
dataclasses-json
ddgs
fastapi
fastapi-cli
groq
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-groq
langchain-openai
langchain-text-splitters
openai
pydantic
python-dotenv
python-multipart
requests
uvicorn
